Implement initial mechanism to read the thread-state from TLS
