+++ pkg/profiler/pyperf/headers.go
@@ -18,5 +18,6 @@ type InterpreterInfo struct {
 	// u64 start_time;
 	// InterpreterAddr uint64
 	ThreadStateAddr      uint64
+	TLSKeyAddr           uint64
 	PyVersionOffsetIndex uint32
 }
